Working with Default Map
